Update.
2000-12-10  Andreas Jaeger  <aj@suse.de>

	* locale/programs/ld-ctype.c (charclass_symbolic_ellipsis): Fixed
	cast as suggested by Jakub Jelinek <jakub@redhat.com>.

2000-12-06  Jim Wilson	<wilson@redhat.com>
2000-12-05  Martin Schwidefsky	<schwidefsky@de.ibm.com>
2000-11-22  Paul Eggert	 <eggert@twinsun.com>
2000-11-28  Masahide Washizawa	<washi@jp.ibm.com>
2000-11-21  Alan Modra	<alan@linuxcare.com.au>
2000-11-21  Martin Schwidefsky	<schwidefsky@de.ibm.com>
	failed.	 Use gettext() not _().
2000-11-09  Martin Schwidefsky	<mschwide@nc.boeblingen.de.ibm.com>

+++ locale/programs/ld-ctype.c
@@ -1482,8 +1482,10 @@ charclass_symbolic_ellipsis (struct linereader *ldfile,
 	  struct charseq *seq;
 	  uint32_t wch;
 
-	  sprintf (tmp, (base == 10 ? "%.*s%0*ld" : "%.*s%0*lX"), cp - last_str,
-		   last_str, (int) (now->val.str.lenmb - (cp - last_str)), from);
+	  sprintf (tmp, (base == 10 ? "%.*s%0*ld" : "%.*s%0*lX"),
+		   cp - last_str, last_str,
+		   (int) (now->val.str.lenmb - (int) (cp - last_str)),
+		   from);
 
 	  get_character (now, charmap, repertoire, &seq, &wch);
